Move `which` to gmx.util
